+++ src/base-element.js
@@ -1,6 +1,9 @@
 // @ts-check
-import { LitElement } from "lit";
+import {
+  LitElement,
+} from "lit";
 import { DefineableMixin } from "web-component-define";
-import { ScopedElementsMixin } from "@open-wc/scoped-elements"
+import { LitScopedElementsMixin } from "web-component-define/exports/lit.js"
 
-export class BaseElement extends ScopedElementsMixin(DefineableMixin(LitElement)) {}
+/** @type HTMLElement */
+export class BaseElement extends DefineableMixin(LitScopedElementsMixin(LitElement)) {}